requirements.txt
fastapi
uvicorn
dashscope
requests
python-multipart
jinja2
aiofiles
pydantic
python-dotenv
